refactor: move AllSomeIter to its own module
