+++ src/ldtk/ldtk_fields.rs
@@ -1,40 +1,11 @@
 //! Contains [`LdtkFields`] trait, providing convenience methods for accessing field instances.
 use crate::ldtk::{
-    EntityInstance, FieldInstance, FieldInstanceEntityReference, FieldValue, Level,
-    TilesetRectangle,
+    all_some_iter::AllSomeIter, EntityInstance, FieldInstance, FieldInstanceEntityReference,
+    FieldValue, Level, TilesetRectangle,
 };
 use bevy::prelude::*;
 use paste::paste;
-use std::{iter::Flatten, slice::Iter};
 use thiserror::Error;
-
-pub struct NotAllSomeError;
-
-pub struct AllSomeIter<'a, T> {
-    flattened: Flatten<Iter<'a, Option<T>>>,
-}
-
-impl<'a, T> Iterator for AllSomeIter<'a, T> {
-    type Item = &'a T;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.flattened.next()
-    }
-}
-
-impl<'a, T> TryFrom<&'a [Option<T>]> for AllSomeIter<'a, T> {
-    type Error = NotAllSomeError;
-
-    fn try_from(value: &'a [Option<T>]) -> Result<Self, Self::Error> {
-        if value.iter().all(|v| v.is_some()) {
-            Ok(AllSomeIter {
-                flattened: value.iter().flatten(),
-            })
-        } else {
-            Err(NotAllSomeError)
-        }
-    }
-}
 
 /// Errors related to the [`LdtkFields`] trait.
 #[derive(Debug, PartialEq, Eq, Error)]
