+++ src/ldtk/mod.rs
@@ -32,6 +32,7 @@ use std::collections::HashMap;
 #[allow(unused_imports)]
 use crate::prelude::LdtkEntity;
 
+pub mod all_some_iter;
 mod color;
 mod field_instance;
 pub mod ldtk_fields;
